Revert "Moved notifications back to top level because it might break some installations"
This reverts commit 0c3229d5ac9129b6560706b85153a94c0585335e.
